handling_files.py and initiating_process.py are created and put necessary codes
