+++ flask_api/app/static/statistics.py
@@ -1,0 +1,20 @@
+def find_average(combined_dict):
+    keys_to_ignore = {"id", "userId", "videoId", "voiceGraphBase64"}
+    keys = [key for key in combined_dict[0].keys() if key not in keys_to_ignore]
+    
+    sums = {key: 0 for key in keys}
+    count = len(combined_dict)
+    
+    for obj in combined_dict:
+        for key in keys:
+            sums[key] += obj[key]
+    
+    averages = {key: round(sums[key] / count, 2) for key in keys}
+
+    average_dict = {
+        "userId": combined_dict[0]["userId"],
+        **averages,
+        "voiceGraphBase64": "iVBORw0KGgoAAAANSUhEUgAADhAA"
+    }
+    
+    return average_dict
